requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-dotenv
pydantic
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
boto3
aiofiles
websockets 
